add some requests and fix minors in controllers

+++ public/FOTH/routes/role.php
@@ -5,11 +5,11 @@ use App\Role;
 
 Route::get('/', 'RoleController@index');
 
-Route::get('/{roleId}', 'RoleController@show');
+Route::get('/{roleId}', 'RoleController@show')->where('roleId', '[0-9]+');
 
 Route::post('/', 'RoleController@create');
 
-Route::put('/{roleId}', 'RoleController@update');
+Route::put('/{roleId}', 'RoleController@update')->where('roleId', '[0-9]+');
 
-Route::delete('/{roleId}','RoleController@destroy');
+Route::delete('/{roleId}','RoleController@destroy')->where('roleId', '[0-9]+');
